// src/types/types.tsx
export enum Types {
 HALTERO = "Haltéro",
 BENCHMARK = "Benchmark",
 GYM = "Gym",
 CARDIO = "Cardio",
}

export enum Haltero {
 CleanAndJerk = "Clean and Jerk",
 Snatch = "Snatch",
 Deadlift = "Deadlift",
 Thruster = "Thruster",
 OverheadSquat = "Overhead Squat",
 HangClean = "Hang Clean",
 PushPress = "Push Press",
 PowerClean = "Power Clean",
}

export enum Benchmark {
 Fran = "Fran",
 Grace = "Grace",
 Helen = "Helen",
 Isabel = "Isabel",
 Diane = "Diane",
 Jackie = "Jackie",
 Cindy = "Cindy",
 Annie = "Annie",
 Murph = "Murph",
}

export enum Gym {
 MuscleUp = "Muscle-Up",
 ToesToBar = "Toes to Bar",
 PullUp = "Pull-Up",
 HandstandPushUp = "Handstand Push-Up",
 PushUp = "Push-Up",
 RingDips = "Ring Dips",
 BarDips = "Bar Dips",
 HandstandWalk = "Handstand Walk",
 LSit = "L-Sit",
 ChestToBar = "Chest to Bar",
}

export enum Cardio {
 Run = "Run",
}